requirements.txt
azure-functions
requests
beautifulsoup4
sendgrid
